Refactor basic auth in vicadm
Removed basic auth handler.  The client just needs it's parameters
checked and, on error, get a header with the appropriate message in
error.  The interface now just checks if the supplied u/p are ok or not
and returns the appropriate fields.
